+++ deduce/tokenizer.py
@@ -4,13 +4,13 @@ from typing import Iterable, Optional
 import docdeid as dd
 import regex
 
-TOKENIZER_PATTERN = regex.compile(r"[a-z]+|\d+|[\n\r]|.(?<! )", flags=re.I | re.M)
+TOKENIZER_PATTERN = regex.compile(r"\w+|[\n\r\t]|.(?<! )", flags=re.I | re.M)
 
 
 class DeduceTokenizer(dd.tokenize.Tokenizer):
     """
-    Tokenizes text, where a token is any sequence of alpha characters (case insensitive), any sequence of numeric
-    characters, a single newline character, or a single special character. It does not include whitespaces as tokens.
+    Tokenizes text, where a token is any sequence of alphanumeric characters (case insensitive), a single
+    newline/tab character, or a single special character. It does not include whitespaces as tokens.
 
     Arguments:
         merge_terms: An iterable of strings that should not be split (i.e. always returned as tokens).